requirements.txt
streamlit==1.29.0
python-docx==1.1.0
pypdf==3.17.0
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
